+++ core/src/main/kotlin/net/corda/core/internal/Emoji.kt
@@ -45,6 +45,8 @@ object Emoji {
     val CODE_FREE: String = codePointsString(0x1F193)
     @JvmStatic
     val CODE_SOON: String = codePointsString(0x1F51C)
+    @JvmStatic
+    val CODE_DEVELOPER: String = codePointsString(0x1F469, 0x200D, 0x1F4BB)
 
 
     /**
@@ -65,6 +67,7 @@ object Emoji {
     val lightBulb: String get() = if (emojiMode.get() != null) "$CODE_LIGHTBULB  " else ""
     val free: String get() = if (emojiMode.get() != null) "$CODE_FREE  " else ""
     val soon: String get() = if (emojiMode.get() != null) "$CODE_SOON  " else ""
+    val developer: String get() = if (emojiMode.get() != null) "$CODE_DEVELOPER  " else ""
 
     // These have old/non-emoji symbols with better platform support.
     val greenTick: String get() = if (emojiMode.get() != null) "$CODE_GREEN_TICK  " else "✓"