+++ node/src/main/kotlin/net/corda/node/internal/AbstractNode.kt
@@ -14,6 +14,7 @@ import net.corda.core.flows.*
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.core.identity.PartyAndCertificate
+import net.corda.core.internal.Emoji
 import net.corda.core.internal.FlowStateMachine
 import net.corda.core.internal.VisibleForTesting
 import net.corda.core.internal.concurrent.map
@@ -173,8 +174,10 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
 
     private fun initCertificate() {
         if (configuration.devMode) {
-            log.warn("Corda node is running in dev mode.")
+            log.warn("Corda node is running in developer mode. This is not suitable for production.")
             configuration.configureWithDevSSLCertificate()
+        } else {
+            log.info("Corda node is running in production mode. If this is a developer environment you can set 'devMode=true' in the node.conf file.")
         }
         validateKeystore()
     }
@@ -199,6 +202,15 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
 
     open fun start(): StartedNode<AbstractNode> {
         check(started == null) { "Node has already been started" }
+        if (configuration.devMode) {
+            Emoji.renderIfSupported {
+                println("""---------------- Attention! -------------------------------
+                    |${Emoji.skullAndCrossbones} Node is running in developer mode! ${Emoji.developer}
+                    |${Emoji.skullAndCrossbones} This is NOT safe for a production deployment.
+                    |-----------------------------------------------------------
+                """.trimMargin())
+            }
+        }
         log.info("Node starting up ...")
         initCertificate()
         initialiseJVMAgents()