+++ node/src/test/kotlin/net/corda/node/services/config/NodeConfigurationImplTest.kt
@@ -1,13 +1,19 @@
 package net.corda.node.services.config
 
+import com.typesafe.config.Config
+import com.typesafe.config.ConfigFactory
 import net.corda.core.internal.div
+import net.corda.core.internal.toPath
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.seconds
 import net.corda.testing.core.ALICE_NAME
 import net.corda.testing.node.MockServices.Companion.makeTestDataSourceProperties
 import net.corda.tools.shell.SSHDConfiguration
+import org.apache.commons.lang3.SystemUtils
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Test
+import java.lang.reflect.Field
+import java.lang.reflect.Modifier
 import java.nio.file.Paths
 import java.util.*
 import kotlin.test.assertFalse
@@ -41,6 +47,70 @@ class NodeConfigurationImplTest {
         assertFalse { testConfiguration.copy(noLocalShell = true).shouldInitCrashShell() }
         assertFalse { testConfiguration.copy(sshd = null).shouldInitCrashShell() }
         assertFalse { testConfiguration.copy(noLocalShell = true, sshd = null).shouldInitCrashShell() }
+    }
+
+    @Test
+    fun `Dev mode is autodetected correctly`() {
+        val os = System.getProperty("os.name")
+
+        setSystemOs("Windows 98")
+        assertTrue(getConfig("test-config-empty.conf").getBoolean("devMode"))
+
+        setSystemOs("Mac Sierra")
+        assertTrue(getConfig("test-config-empty.conf").getBoolean("devMode"))
+
+        setSystemOs("Windows server 2008")
+        assertFalse(getConfig("test-config-empty.conf").getBoolean("devMode"))
+
+        setSystemOs("Linux")
+        assertFalse(getConfig("test-config-empty.conf").getBoolean("devMode"))
+
+        setSystemOs(os)
+    }
+
+    private fun setSystemOs(os: String) {
+        val fieldOsName = SystemUtils::class.java.getDeclaredField("OS_NAME")
+        val fieldOsMac = SystemUtils::class.java.getDeclaredField("IS_OS_MAC")
+        val fieldOsWin = SystemUtils::class.java.getDeclaredField("IS_OS_WINDOWS")
+        val modifiersField = Field::class.java.getDeclaredField("modifiers")
+        modifiersField.isAccessible = true
+        modifiersField.setInt(fieldOsName, fieldOsName.getModifiers() and Modifier.FINAL.inv())
+        modifiersField.setInt(fieldOsMac, fieldOsMac.getModifiers() and Modifier.FINAL.inv())
+        modifiersField.setInt(fieldOsWin, fieldOsWin.getModifiers() and Modifier.FINAL.inv())
+        fieldOsName.set(null, os)
+        fieldOsMac.set(null, os.contains("Mac"))
+        fieldOsWin.set(null, os.contains("Windows"))
+        System.setProperty("os.name", os)
+    }
+
+    @Test
+    fun `Dev mode is read from the config over the autodetect logic`() {
+        assertTrue(getConfig("test-config-DevMode.conf").getBoolean("devMode"))
+        assertFalse(getConfig("test-config-noDevMode.conf").getBoolean("devMode"))
+    }
+
+    @Test
+    fun `Dev mode is true if overriden`() {
+        assertTrue(getConfig("test-config-DevMode.conf", ConfigFactory.parseMap(mapOf("devMode" to true))).getBoolean("devMode"))
+        assertTrue(getConfig("test-config-noDevMode.conf", ConfigFactory.parseMap(mapOf("devMode" to true))).getBoolean("devMode"))
+        assertTrue(getConfig("test-config-empty.conf", ConfigFactory.parseMap(mapOf("devMode" to true))).getBoolean("devMode"))
+    }
+
+    @Test
+    fun `Dev mode is false if overriden`() {
+        assertFalse(getConfig("test-config-DevMode.conf", ConfigFactory.parseMap(mapOf("devMode" to false))).getBoolean("devMode"))
+        assertFalse(getConfig("test-config-noDevMode.conf", ConfigFactory.parseMap(mapOf("devMode" to false))).getBoolean("devMode"))
+        assertFalse(getConfig("test-config-empty.conf", ConfigFactory.parseMap(mapOf("devMode" to false))).getBoolean("devMode"))
+    }
+
+    private fun getConfig(cfgName: String, overrides: Config = ConfigFactory.empty()): Config {
+        val path = this::class.java.classLoader.getResource(cfgName).toPath()
+        val cfg = ConfigHelper.loadConfig(
+                baseDirectory = path.parent,
+                configFile = path,
+                configOverrides = overrides
+        )
+        return cfg
     }
 
     private fun configDebugOptions(devMode: Boolean, devModeOptions: DevModeOptions?): NodeConfiguration {